Implement fast path
The bot currently takes many hours to run each time, with considerable network and CPU load. We can optimize it to skip over projects where the best version on wikidata and the latest release on github are the same.

The drawback is that we don't backfill projects if the latest release was manually added, but that's acceptable for the significantly improved performance. There's also some risk that the logic is wrong and we're skipping over projects we should update.